* Correção ao pegar carrinho em npc. - Ao relogar o carrinho não estava sendo salvo. * Adicionado Cavaleiro Rúnico no check de montaria.

+++ brathena/db/pt-BR/pre-renovacao.sql
@@ -29583,7 +29583,7 @@ REPLACE INTO `statpoint_db` VALUES(7185);
 
 CREATE TABLE IF NOT EXISTS `sc_config` (
   `SC_NAME` varchar(70) NOT NULL,
-  `flag` tinyint NOT NULL,
+  `flag` smallint unsigned NOT NULL,
   PRIMARY KEY (`SC_NAME`)
 ) ENGINE=MyISAM DEFAULT CHARSET=latin1;
